backend_filmes/pyproject.toml
[tool.poetry]
name = "backend_filmes"
version = "0.1.0"
description = ""
authors = ["Adarah <lucasyharada@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"
isort = "^5.9.1"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
